I making a Weather App - 1
Today I started a weather app. I'm trying to do it with the react knowledge I've learned so far. For example, I design components to be reused, use useEffect and useState, and also fetch data with axios. For today, I have finished half of both the interface and the fetch processes.